fix: wire htr-xnglo into the real plugin project + CI
- restored the official PluginInterface.h (my earlier hand-written
  version in xnglonpp-ext was a simplified guess and used wchar_t
  differently / older Sci_TextRange API -- this repo's real
  Notepad_plus_msgs.h/Scintilla.h are authoritative, so use them instead
  of my own)
- PluginDefinition.cpp/dllmain.cpp rewritten against the actual modern
  API: wchar_t plugin/menu names, SCI_GETTEXTRANGEFULL + Sci_TextRangeFull
  (SCI_GETTEXTRANGE/Sci_TextRange are deprecated in this Scintilla.h)
- vcxproj: removed the demo's NppPluginDemo.cpp/.rc and DockingFeature
  files (unused by this plugin) from the build, added dllmain.cpp +
  xnglo_core/*.cpp/*.h; set TargetName=htr-xnglo so the built DLL is
  literally htr-xnglo.dll/.lib
- CI_build.yml (already present in this template, MSBuild across
  Release/Debug x x64/Win32/ARM64) artifact paths updated to match

Not compiled with a real MSVC anywhere -- this is a careful manual
review against the template's actual headers, not a verified build.
The pushed CI run is the first real compile attempt; see its log for
whatever this review missed.

+++ src/xnglo_core/core.cpp
@@ -182,7 +182,7 @@ std::string to_xi38(const std::string& utf8_input) {
   for (char32_t cp : cps) {
     const ScriptTable* t = table_for(cp);
     if (t) {
-      out += t->map[cp - t->base];
+      out += t->map[static_cast<size_t>(cp) - static_cast<size_t>(t->base)];
     } else {
       utf8_append(out, cp);
     }
